+++ test/Spec.hs
@@ -61,11 +61,11 @@ main = do
         tid <- TID.genTypeID ""
         getPrefix tid `shouldBe` ""
       it "can parse TypeID from String" do
-        case TID.parseString "mmzk_00041061050r3gg28a1c60t3gf" of
+        case string2ID @TypeID "mmzk_00041061050r3gg28a1c60t3gf" of
           Left err  -> expectationFailure $ "Parse error: " ++ show err
           Right tid -> getPrefix tid `shouldBe` "mmzk"
       it "has the correct nil" do
-        Right TID.nilTypeID `shouldBe` TID.parseString "00000000000000000000000000"
+        Right TID.nilTypeID `shouldBe` string2ID "00000000000000000000000000"
       it "can generate in batch with same timestamp and in ascending order" do
         tids <- TID.genTypeIDs "mmzk" 1526
         all ((== "mmzk") . getPrefix) tids `shouldBe` True
@@ -96,9 +96,9 @@ main = do
                             , ("wrong_alphabet", "ooooooiiiiiiuuuuuuulllllll") ]
       describe "can detect invalid suffix" do
         forM_ invalidSuffixes \(reason, suffix) -> it reason do
-          case TID.parseString suffix of
+          case string2ID @TypeID suffix of
             Left _    -> pure ()
-            Right tid -> expectationFailure $ "Parsed TypeID: " ++ TID.toString tid
+            Right tid -> expectationFailure $ "Parsed TypeID: " ++ id2String tid
 
     describe "Parse special values" do
       let specialValues = [ ("nil", "00000000000000000000000000", "00000000-0000-0000-0000-000000000000")
@@ -107,7 +107,7 @@ main = do
                           , ("sixteen", "0000000000000000000000000g", "00000000-0000-0000-0000-000000000010")
                           , ("thirty-two", "00000000000000000000000010", "00000000-0000-0000-0000-000000000020") ]
       forM_ specialValues \(reason, tid, uuid) -> it reason do
-        case TID.parseString tid of
+        case string2ID @TypeID tid of
           Left err  -> expectationFailure $ "Parse error: " ++ show err
           Right tid -> V7.toString (KID.getUUID tid) `shouldBe` uuid
 
@@ -140,7 +140,7 @@ main = do
         forM_ invalid \(TestData name tid _ _) -> it name do 
           case decode (fromString $ show tid) :: Maybe TypeID of
             Nothing  -> pure ()
-            Just tid -> expectationFailure $ "Parsed TypeID: " ++ TID.toString tid
+            Just tid -> expectationFailure $ "Parsed TypeID: " ++ id2String tid
       describe "Invalid JSON key" do
         forM_ invalid \(TestData name tid _ _) -> it name do 
           case decode (fromString $ "{" ++ show tid ++ ":" ++ "114514" ++ "}") :: Maybe (Map TypeID Int) of
@@ -149,13 +149,13 @@ main = do
 
     describe "Test invalid.json" do
       forM_ invalid \(TestData name tid _ _) -> it name do 
-        case TID.parseString tid of
+        case string2ID @TypeID tid of
           Left _    -> pure ()
-          Right tid -> expectationFailure $ "Parsed TypeID: " ++ TID.toString tid
+          Right tid -> expectationFailure $ "Parsed TypeID: " ++ id2String tid
 
     describe "Test valid.json (TypeID as literal)" do
       forM_ valid \(TestData name tid (Just prefix) (Just uuid)) -> it name do
-        case TID.parseString tid of
+        case string2ID @TypeID tid of
           Left err  -> expectationFailure $ "Parse error: " ++ show err
           Right tid -> do
             getPrefix tid `shouldBe` prefix
@@ -174,15 +174,15 @@ main = do
         kid <- KID.genKindID @""
         getPrefix kid `shouldBe` ""
       it "can parse TypeID from String" do
-        case KID.parseString @"mmzk" "mmzk_00041061050r3gg28a1c60t3gf" of
+        case string2ID @(KindID "mmzk") "mmzk_00041061050r3gg28a1c60t3gf" of
           Left err  -> expectationFailure $ "Parse error: " ++ show err
           Right kid -> getPrefix kid `shouldBe` "mmzk"
       it "cannot parse TypeID into wrong prefix" do
-        case KID.parseString @"foo" "mmzk_00041061050r3gg28a1c60t3gf" of
+        case string2ID @(KindID "foo") "mmzk_00041061050r3gg28a1c60t3gf" of
           Left err  -> pure ()
-          Right kid -> expectationFailure $ "Parsed TypeID: " ++ KID.toString kid
+          Right kid -> expectationFailure $ "Parsed TypeID: " ++ id2String kid
       it "has the correct nil" do
-        Right KID.nilKindID `shouldBe` KID.parseString @"" "00000000000000000000000000"
+        Right KID.nilKindID `shouldBe` string2ID "00000000000000000000000000"
       it "can generate in batch with same timestamp and in ascending order" do
         kids <- KID.genKindIDs @"mmzk" 1526
         all ((== "mmzk") . getPrefix) kids `shouldBe` True
@@ -195,13 +195,13 @@ main = do
           kid <- KID.genKindID @'Post
           getPrefix kid `shouldBe` "post"
       it "can parse TypeID from String" do
-        case KID.parseString @'User "user_00041061050r3gg28a1c60t3gf" of
+        case string2ID @(KindID User) "user_00041061050r3gg28a1c60t3gf" of
           Left err  -> expectationFailure $ "Parse error: " ++ show err
           Right kid -> getPrefix kid `shouldBe` "user"
       it "cannot parse TypeID into wrong prefix" do
-        case KID.parseString @'Comment "user_00041061050r3gg28a1c60t3gf" of
+        case string2ID @(KindID Comment) "user_00041061050r3gg28a1c60t3gf" of
           Left err  -> pure ()
-          Right kid -> expectationFailure $ "Parsed TypeID: " ++ KID.toString kid
+          Right kid -> expectationFailure $ "Parsed TypeID: " ++ id2String kid
       it "can generate in batch with same timestamp and in ascending order" do
         kids <- KID.genKindIDs @'Comment 1526
         all ((== "comment") . getPrefix) kids `shouldBe` True